Allow user to add new leads.
* New lead is a 'route' and not a 'resource' because it does not need to
load an existing model in the system.
* By keeping the attributes in 'fields' and waiting until the action to
create a record is called, the record won’t appear in the list until the
user has chosen to create it. This property is set to an empty object in
the route so that it’s reset every time the user visit the new lead route.
* The template goes in 'templates/leads/' because it’s a 'route' nested
under a 'resource' named 'leads'.
* Perform custom validations on first and last name.